[SEARCH-1430] Moving to static approach
