+++ alfresco-search/src/test/java/org/alfresco/solr/AdminHandlerDistributedTest.java
@@ -26,14 +26,12 @@ import org.apache.solr.core.SolrCore;
 import org.apache.solr.response.SolrQueryResponse;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
-import org.junit.Rule;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.lang.invoke.MethodHandles;
 import java.util.Properties;
-import java.util.concurrent.TimeUnit;
 
 import static org.alfresco.repo.index.shard.ShardMethodEnum.*;
 import static org.alfresco.solr.AlfrescoSolrUtils.*;
@@ -45,7 +43,7 @@ import static org.alfresco.solr.AlfrescoSolrUtils.*;
  */
 @SolrTestCaseJ4.SuppressSSL
 @LuceneTestCase.SuppressCodecs({"Appending","Lucene3x","Lucene40","Lucene41","Lucene42","Lucene43", "Lucene44", "Lucene45","Lucene46","Lucene47","Lucene48","Lucene49"})
-public class AdminHandlerDistributedTest extends AbstractAlfrescoDistributedTestStatic
+public class AdminHandlerDistributedTest extends AbstractAlfrescoDistributedTest
 {
     private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
     final String JETTY_SERVER_ID = this.getClass().getSimpleName();
