+++ alfresco-search/src/test/java/org/alfresco/solr/tracker/DistributedAlfrescoSolrJsonTest.java
@@ -18,10 +18,13 @@
  */
 package org.alfresco.solr.tracker;
 
-import org.alfresco.solr.AbstractAlfrescoDistributedTest;
 import org.alfresco.repo.search.adaptor.lucene.QueryConstants;
-import org.alfresco.solr.AbstractAlfrescoDistributedTestStatic;
-import org.alfresco.solr.client.*;
+import org.alfresco.solr.AbstractAlfrescoDistributedTest;
+import org.alfresco.solr.client.Acl;
+import org.alfresco.solr.client.AclChangeSet;
+import org.alfresco.solr.client.Node;
+import org.alfresco.solr.client.NodeMetaData;
+import org.alfresco.solr.client.Transaction;
 import org.apache.lucene.index.Term;
 import org.apache.lucene.search.BooleanClause;
 import org.apache.lucene.search.BooleanQuery;
@@ -32,22 +35,22 @@ import org.apache.solr.SolrTestCaseJ4;
 import org.apache.solr.client.solrj.response.QueryResponse;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
-import org.junit.Rule;
 import org.junit.Test;
 
-import static org.alfresco.solr.AlfrescoSolrUtils.*;
-import static org.alfresco.solr.AlfrescoSolrUtils.getAclReaders;
-import static org.alfresco.solr.AlfrescoSolrUtils.list;
-
-import java.util.List;
 import java.util.ArrayList;
+import java.util.List;
+
+import static org.alfresco.solr.AlfrescoSolrUtils.TestActChanges;
+import static org.alfresco.solr.AlfrescoSolrUtils.getNode;
+import static org.alfresco.solr.AlfrescoSolrUtils.getNodeMetaData;
+import static org.alfresco.solr.AlfrescoSolrUtils.getTransaction;
 
 /**
  * @author Joel
  */
 @SolrTestCaseJ4.SuppressSSL
 @LuceneTestCase.SuppressCodecs({"Appending","Lucene3x","Lucene40","Lucene41","Lucene42","Lucene43", "Lucene44", "Lucene45","Lucene46","Lucene47","Lucene48","Lucene49"})
-public class DistributedAlfrescoSolrJsonTest extends AbstractAlfrescoDistributedTestStatic
+public class DistributedAlfrescoSolrJsonTest extends AbstractAlfrescoDistributedTest
 {
     @BeforeClass
     private static void initData() throws Throwable
